Add udp send benchmark, but caveats
this only sends, because receiving seems to cause blocking (unverified)
in addition, the kernel symbols and timings are all weird, and sending
seems to be unnaturally slow (like, sending 100 messages takes on the
order of seconds, when it really shouldn't be that slow). this seems to
happen even on a stand-alone non-benchmark binary.